# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "API for an employee management system to determine employees that\
    \ are no longer working on active projects and to understand their qualifications\
    \ and expertise for suitability in other projects"
  version: "1.0.0"
  title: "UBahn API"
...
schemes:
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /users/{userId}/achievements:
    get:
      tags:
      - "Achievements"
      description: "Get Achievements for given user id.\n\n**Security** - Note that\
        \ for non-admin users, this endpoint will only return entities that\nthe user\
        \ has created.        \n"
      operationId: "usersUserIdAchievementsGET"
      parameters:
      - name: "userId"
        in: "path"
        description: "The user id"
        required: true
        type: "string"
        format: "UUID"
      - name: "achievementsproviderName"
        in: "query"
        description: "The achievement provider name"
        required: false
        type: "string"
      responses:
        "200":
          description: "OK - the request was successful"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Achievement"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "404":
          $ref: "#/definitions/NotFound"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "Achievements"
    head:
      tags:
      - "Achievements"
      description: "Get Achievements for given user id, but only header information\
        \ is returned.\n\n**Security** - Note that for non-admin users, this endpoint\
        \ will only return entities that\nthe user has created.        \n"
      operationId: "usersUserIdAchievementsHEAD"
      parameters:
      - name: "userId"
        in: "path"
        description: "The user id"
        required: true
        type: "string"
        format: "UUID"
      - name: "achievementsproviderName"
        in: "query"
        description: "The achievement provider name"
        required: false
        type: "string"
      responses:
        "200":
          description: "OK - the request was successful"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "404":
          $ref: "#/definitions/NotFound"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "Achievements"
    post:
      tags:
      - "Achievements"
      description: "Create a new Achievement.\n\n**Security** - This endpoint is accessible\
        \ by all authenticated users.        \n"
      operationId: "usersUserIdAchievementsPOST"
      parameters:
      - name: "userId"
        in: "path"
        description: "The user id"
        required: true
        type: "string"
        format: "UUID"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/AchievementRequestBody"
      responses:
        "200":
          description: "OK - the request was successful"
          schema:
            $ref: "#/definitions/Achievement"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "409":
          $ref: "#/definitions/Conflict"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "Achievements"
  /users/{userId}/achievements/{achievementsProviderId}:
    get:
      tags:
      - "Achievements"
      description: "Get Achievements for given user id and provider id.\n\n**Security**\
        \ - Note that for non-admin users, this endpoint will only return entities\
        \ that\nthe user has created.\n"
      operationId: "usersUserIdAchievementsAchievementsProviderIdGET"
      parameters:
      - name: "userId"
        in: "path"
        description: "The user id"
        required: true
        type: "string"
        format: "UUID"
      - name: "achievementsProviderId"
        in: "path"
        description: "The provider id"
        required: true
        type: "string"
        format: "UUID"
      responses:
        "200":
          description: "OK - the request was successful"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Achievement"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "404":
          $ref: "#/definitions/NotFound"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "Achievements"
    head:
      tags:
      - "Achievements"
      description: "Get Achievements for given user id and provider id, but only header\
        \ information is returned.\n\n**Security** - Note that for non-admin users,\
        \ this endpoint will only return entities that\nthe user has created.    \
        \    \n"
      operationId: "usersUserIdAchievementsAchievementsProviderIdHEAD"
      parameters:
      - name: "userId"
        in: "path"
        description: "The user id"
        required: true
        type: "string"
        format: "UUID"
      - name: "achievementsProviderId"
        in: "path"
        description: "The provider id"
        required: true
        type: "string"
        format: "UUID"
      responses:
        "200":
          description: "OK - the request was successful"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "404":
          $ref: "#/definitions/NotFound"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "Achievements"
    delete:
      tags:
      - "Achievements"
      description: "Remove an existing Achievement with given userId and achievement\
        \ providerId.\n\n**Security** - Note that this endpoint is only available\
        \ for admin users.  \n"
      operationId: "usersUserIdAchievementsAchievementsProviderIdDELETE"
      parameters:
      - name: "userId"
        in: "path"
        description: "The user id"
        required: true
        type: "string"
        format: "UUID"
      - name: "achievementsProviderId"
        in: "path"
        description: "The provider id"
        required: true
        type: "string"
        format: "UUID"
      responses:
        "204":
          description: "OK - the request was successful"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "404":
          $ref: "#/definitions/NotFound"
        "409":
          $ref: "#/definitions/Conflict"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "Achievements"
    patch:
      tags:
      - "Achievements"
      description: "Update an existing Achievement with given userId and achievement\
        \ providerId.  Only the fields in the request body are updated.\n\n**Security**\
        \ - Note that for non-admin users, this endpoint will only allow updates on\
        \ entities that the\ncalling user has created.\n"
      operationId: "usersUserIdAchievementsAchievementsProviderIdPATCH"
      parameters:
      - name: "userId"
        in: "path"
        description: "The user id"
        required: true
        type: "string"
        format: "UUID"
      - name: "achievementsProviderId"
        in: "path"
        description: "The provider id"
        required: true
        type: "string"
        format: "UUID"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/AchievementUpdateRequestBody"
      responses:
        "200":
          description: "OK - the request was successful"
          schema:
            $ref: "#/definitions/Achievement"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "404":
          $ref: "#/definitions/NotFound"
        "409":
          $ref: "#/definitions/Conflict"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "Achievements"
  /achievementsProviders:
    get:
      tags:
      - "Achievements Provider"
      description: "Search Achievements Provider in the application.  If no results,\
        \ then empty array is returned.\n\n**Security** - Note that for non-admin\
        \ users, this endpoint will only return entities that\nthe user has created.\n"
      operationId: "achievementsProvidersGET"
      parameters:
      - name: "name"
        in: "query"
        description: "Filter by provider name"
        required: false
        type: "string"
      responses:
        "200":
          description: "OK - the request was successful"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/AchievementsProvider"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "AchievementsProvider"
    head:
      tags:
      - "Achievements Provider"
      description: "Retrieve header information for a search operation on achivements\
        \ provider in the application.\n\n**Security** - Note that for non-admin users,\
        \ this endpoint will only return entities that\nthe user has created.\n"
      operationId: "achievementsProvidersHEAD"
      parameters:
      - name: "name"
        in: "query"
        description: "Filter by provider name"
        required: false
        type: "string"
      responses:
        "200":
          description: "Success response"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "AchievementsProvider"
    post:
      tags:
      - "Achievements Provider"
      description: "Create a new Achievements Provider.\n\n**Security** - This endpoint\
        \ is accessible by all authenticated users.        \n"
      operationId: "achievementsProvidersPOST"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/NameRequestBody"
      responses:
        "200":
          description: "OK - the request was successful"
          schema:
            $ref: "#/definitions/AchievementsProvider"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "409":
          $ref: "#/definitions/Conflict"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "AchievementsProvider"
  /achievementsProviders/{providerId}:
    get:
      tags:
      - "Achievements Provider"
      description: "Get achievements provider with given id.\n\n**Security** - Note\
        \ that for non-admin users, this endpoint will only return entities that\n\
        the user has created.        \n"
      operationId: "achievementsProvidersProviderIdGET"
      parameters:
      - name: "providerId"
        in: "path"
        description: "The provider id"
        required: true
        type: "string"
        format: "UUID"
      responses:
        "200":
          description: "OK - the request was successful"
          schema:
            $ref: "#/definitions/AchievementsProvider"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "404":
          $ref: "#/definitions/NotFound"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "AchievementsProvider"
    head:
      tags:
      - "Achievements Provider"
      description: "Get achivements provider with given id, but only header information\
        \ is returned.\n\n**Security** - Note that for non-admin users, this endpoint\
        \ will only return entities that\nthe user has created.        \n"
      operationId: "achievementsProvidersProviderIdHEAD"
      parameters:
      - name: "providerId"
        in: "path"
        description: "The provider id"
        required: true
        type: "string"
        format: "UUID"
      responses:
        "200":
          description: "OK - the request was successful"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "404":
          $ref: "#/definitions/NotFound"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "AchievementsProvider"
    delete:
      tags:
      - "Achievements Provider"
      description: "Remove an existing achiements provider with given id.\n\n**Security**\
        \ - Note that this endpoint is only available for admin users.  \n"
      operationId: "achievementsProvidersProviderIdDELETE"
      parameters:
      - name: "providerId"
        in: "path"
        description: "The provider id"
        required: true
        type: "string"
        format: "UUID"
      responses:
        "204":
          description: "OK - the request was successful"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "404":
          $ref: "#/definitions/NotFound"
        "409":
          $ref: "#/definitions/Conflict"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "AchievementsProvider"
    patch:
      tags:
      - "Achievements Provider"
      description: "Update an existing achivements provider with given id.\n\n**Security**\
        \ - Note that for non-admin users, this endpoint will only allow updates on\
        \ entities that the\ncalling user has created.\n"
      operationId: "achievementsProvidersProviderIdPATCH"
      parameters:
      - name: "providerId"
        in: "path"
        description: "The provider id"
        required: true
        type: "string"
        format: "UUID"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/NameRequestBody"
      responses:
        "200":
          description: "OK - the request was successful"
          schema:
            $ref: "#/definitions/AchievementsProvider"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "404":
          $ref: "#/definitions/NotFound"
        "409":
          $ref: "#/definitions/Conflict"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "AchievementsProvider"
...
securityDefinitions:
  Bearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
  AuditFields:
    type: "object"
    required:
    - "created"
    - "createdBy"
    - "updated"
    - "updatedBy"
    properties:
      created:
        type: "string"
        format: "date-time"
        description: "When the entity was created."
      updated:
        type: "string"
        format: "date-time"
        description: "When the entity was updated."
      createdBy:
        type: "string"
        format: "UUID"
        description: "Creator of the entity."
      updatedBy:
        type: "string"
        format: "UUID"
        description: "User that last updated the entity."
    description: "Describes the audit fields that are present in all the models in\
      \ this API."
  Achievement:
    allOf:
    - type: "object"
      required:
      - "achievementsProviderId"
      - "certifierId"
      - "name"
      - "uri"
      - "userId"
      properties:
        userId:
          type: "string"
          format: "UUID"
          description: "The id of user that this Achievement belongs to."
        achievementsProviderId:
          type: "string"
          format: "UUID"
          description: "The id of achievements provider for this Achievement."
        name:
          type: "string"
          description: "Name of achievement."
        uri:
          type: "string"
          description: "Uri of achievement"
        certifierId:
          type: "string"
          description: "Id of certifier"
        certifiedDate:
          type: "string"
          format: "date-time"
          description: "The date when certification occurred."
    - $ref: "#/definitions/AuditFields"
  AchievementRequestBody:
    allOf:
    - type: "object"
      properties:
        achievementsProviderId:
          type: "string"
          format: "UUID"
          description: "The id of provider for this Achievement."
    - $ref: "#/definitions/AchievementUpdateRequestBody"
  AchievementUpdateRequestBody:
    type: "object"
    properties:
      name:
        type: "string"
        description: "Name of Achievement"
      uri:
        type: "string"
        description: "Uri of Achievement"
      certifierId:
        type: "string"
        description: "Id of certifier"
      certifiedDate:
        type: "string"
        format: "date-time"
        description: "The date when certification occurred."
    example:
      certifierId: "certifierId"
      certifiedDate: "2000-01-23T04:56:07.000+00:00"
      name: "name"
      uri: "uri"
  AchievementsProvider:
    allOf:
    - type: "object"
      required:
      - "id"
      - "name"
      properties:
        id:
          type: "string"
          format: "UUID"
          description: "The id of the provider."
        name:
          type: "string"
          description: "The name of the provider."
    - $ref: "#/definitions/AuditFields"
...
  NameRequestBody:
    type: "object"
    required:
    - "name"
    properties:
      name:
        type: "string"
        description: "Name of the entity"
    description: "Simple request body containing the name of the entity."
    example:
      name: "name"
...
  Unauthorized:
    type: "object"
    properties:
      message:
        type: "string"
        example: "Unable to authenticate the user."
        description: "The unauthorized error message."
    description: "The unauthorized error entity."
  NotFound:
    type: "object"
    properties:
      message:
        type: "string"
        example: "A resource with the name could not be found."
        description: "The not found error message."
    description: "The not found error entity."
  ServerError:
    type: "object"
    properties:
      message:
        type: "string"
        example: "Something went wrong while processing your request. We�re sorry\
          \ for the trouble. We�ve been notified of the error and will correct it\
          \ as soon as possible. Please try your request again in a moment."
        description: "The server error message."
    description: "The server error entity."
  BadRequest:
    type: "object"
    properties:
      message:
        type: "string"
        example: "Invalid input."
        description: "The bad request error message."
    description: "The bad request error entity."
  Forbidden:
    type: "object"
    properties:
      message:
        type: "string"
        example: "You are not allowed to access the request."
        description: "The forbidden error message."
    description: "The permission error entity."
  Conflict:
    type: "object"
    required:
    - "message"
    properties:
      message:
        type: "string"
        example: "Creating a resource with a name already exists."
        description: "The conflict error message."
    description: "The conflict error entity."